+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.5.0] - 2026.01.21
+
+* (*Breaking) `isErasing` is now `ErasingBehavior`. Legacy `isErasing: true` is equivalent to `ErasingBehavior.pixel`.
+* Customizable `intersectionDetector` for intersection detection when `ErasingBehaviro.stroke`.
+
 ## [0.4.3] - 2026.01.21
 
 * Updated README.md